feat: 댓글 전역 상태(리덕스) 에서 지역 상태로 변경
