+++ src/component/crudcomponent/CommentSection.jsx
@@ -1,18 +1,35 @@
-import { addDoc, collection } from 'firebase/firestore';
+import { addDoc, collection, getDocs, query, where } from 'firebase/firestore';
 import { db } from '../../firebase';
-import React, { useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import React, { useEffect, useState } from 'react';
+import { useSelector } from 'react-redux';
 import styled from 'styled-components';
-import { addComment } from '../../redux/modules/comment';
 import CommentList from './CommentList';
 import { BtnBlackBg } from 'styles/SharedStyle';
 import { useParams } from 'react-router-dom';
 
 export default function CommentSection() {
   const { user_id, user_img, nickname } = useSelector((state) => state.user.nowUser);
-  const dispatch = useDispatch();
   const params = useParams();
   const [contents, setContents] = useState('');
+  const [commentData, setCommentData] = useState();
+
+  //댓글 데이터 가져오기
+  useEffect(() => {
+    const fetchCommentData = async () => {
+      const q = query(collection(db, 'comments'), where('board_id', '==', params.id));
+      const querySnapshot = await getDocs(q);
+
+      const initialData = [];
+
+      querySnapshot.forEach((doc) => {
+        initialData.push({ id: doc.id, ...doc.data() });
+      });
+
+      setCommentData([...initialData]);
+    };
+
+    fetchCommentData();
+  }, []);
 
   const handleChange = (e) => {
     setContents(e.target.value);
@@ -41,12 +58,15 @@ export default function CommentSection() {
         regDate,
         nickname,
         user_id,
-        // board_id: boardTestData[0].id //임시로 파이어베이스 첫 번째 데이터의 id로 부여함. 나중에 useParam으로 id 가져와서 하는 등 변경 필요
         board_id: params.id
       };
       const collectionRef = collection(db, 'comments');
       const ref = await addDoc(collectionRef, newComment);
-      dispatch(addComment({ ...newComment, id: ref.id }));
+
+      newComment.id = ref.id;
+      setCommentData((prev) => {
+        return [newComment, ...prev];
+      });
       alert('등록이 완료되었습니다.');
       e.target.reset();
     } catch (error) {
@@ -68,7 +88,7 @@ export default function CommentSection() {
           <BtnBlackBg>등록</BtnBlackBg>
         </BtnBox>
       </CommentFormWrap>
-      <CommentList />
+      <CommentList commentData={commentData} setCommentData={setCommentData} />
     </>
   );
 }
